ViberToPDF_Converter 1
Author: Jake Kroon
Revison Number: 1

The aim is to read a Viber backup and convert the data into a readable
PDF, for book keeping purposes. It also makes it more readable to the
average person.

On tested data this works, I just need to find a way to neaten up the
output. For example, words get cut off, I need to read the to the next
space then add a new line, things like this.

General presentation could be vastly improved.

Also structure of class could be nicer, needs work.